requirements.txt
seaborn
slackclient
tensorflow==1.2.0
numpy==1.17.3
torch==1.4.0
torchvision==0.5.0
tqdm==4.31.1
pandas==0.24.2
matplotlib==2.2.3
mxnet==1.5.1.post0
mxnet-cu101==1.5.1.post0
Pillow==5.1.0
requests==2.21.0
decorator==4.3.0
attrs==18.1.0
antlr4-python3-runtime==4.7.2
psutil==5.6.3
